+++ include/Airport/Apple80211.h
@@ -15,8 +15,16 @@
 
 #include "apple_private_spi.h"
 #include "debug.h"
-#include "IO80211WorkQueue.h"
 #include "IO80211WorkLoop.h"
+#ifdef IO80211FAMILY_V2
+#include "IO80211WorkQueue.h"
+#include "IO80211ControllerV2.h"
+#include "IO80211InfraInterface.h"
+#include "IO80211InfraProtocol.h"
+#include "IOSkywalkPacketBufferPool.h"
+#include "IOSkywalkLegacyEthernetInterface.h"
+#include "IO80211SkywalkInterface.h"
+#else
 #include "IO80211Controller.h"
 #include "IO80211Interface.h"
 #include "IO80211VirtualInterface.h"
@@ -24,11 +32,6 @@
 #if __IO80211_TARGET >= __MAC_10_15
 #include "IO80211SkywalkInterface.h"
 #endif
-#if __IO80211_TARGET >= __MAC_14_0
-#include "IO80211InfraInterface.h"
-#include "IO80211InfraProtocol.h"
-#include "IOSkywalkPacketBufferPool.h"
-#include "IOSkywalkLegacyEthernetInterface.h"
-#endif
+#endif /* IO80211FAMILY_V2 */
 
 #endif /* Apple80211_h */
