+++ include/Airport/IO80211Controller.h
@@ -30,11 +30,6 @@
 #include "apple80211_ioctl.h"
 #include "IO80211SkywalkInterface.h"
 #include "IO80211WorkLoop.h"
-#include "IO80211WorkQueue.h"
-#include "CCStream.h"
-#include "CCDataPipe.h"
-#include "CCLogPipe.h"
-#include "CCLogStream.h"
 
 #define AUTH_TIMEOUT            15    // seconds
 
@@ -76,6 +71,10 @@ typedef enum IO80211FeatureCode IO80211FeatureCode;
 
 class IOSkywalkInterface;
 class IO80211ScanManager;
+enum CCStreamLogLevel
+{
+    LEVEL_1,
+};
 
 enum scanSource
 {
@@ -131,133 +130,6 @@ struct apple80211_lteCoex_report;
 typedef IOReturn (*IOCTL_FUNC)(IO80211Controller*, IO80211Interface*, IO80211VirtualInterface*, apple80211req*, bool);
 extern IOCTL_FUNC gGetHandlerTable[];
 extern IOCTL_FUNC gSetHandlerTable[];
-
-#ifdef IO80211FAMILY_V2
-
-class IO80211InterfaceAVCAdvisory;
-
-class IO80211Controller : public IOEthernetController {
-    OSDeclareAbstractStructors(IO80211Controller)
-
-public:
-
-    virtual void free() APPLE_KEXT_OVERRIDE;
-    virtual bool init(OSDictionary *) APPLE_KEXT_OVERRIDE;
-    virtual IOReturn configureReport(IOReportChannelList *,UInt,void *,void *) APPLE_KEXT_OVERRIDE;
-    virtual IOReturn updateReport(IOReportChannelList *,UInt,void *,void *) APPLE_KEXT_OVERRIDE;
-    virtual bool start(IOService *) APPLE_KEXT_OVERRIDE;
-    virtual void stop(IOService *) APPLE_KEXT_OVERRIDE;
-    virtual IOWorkLoop* getWorkLoop(void) const APPLE_KEXT_OVERRIDE;
-    virtual const char* stringFromReturn(int) APPLE_KEXT_OVERRIDE;
-    virtual int errnoFromReturn(int) APPLE_KEXT_OVERRIDE;
-    virtual UInt32 getFeatures() const APPLE_KEXT_OVERRIDE;
-    virtual const OSString * newVendorString() const APPLE_KEXT_OVERRIDE;
-    virtual const OSString * newModelString() const APPLE_KEXT_OVERRIDE;
-    virtual bool createWorkLoop() APPLE_KEXT_OVERRIDE;
-    virtual IOReturn getHardwareAddress(IOEthernetAddress *) APPLE_KEXT_OVERRIDE;
-    virtual IOReturn setHardwareAddress(const IOEthernetAddress * addrP) APPLE_KEXT_OVERRIDE;
-    virtual IOReturn setMulticastMode(bool active) APPLE_KEXT_OVERRIDE;
-    virtual IOReturn setPromiscuousMode(bool active) APPLE_KEXT_OVERRIDE;
-    virtual bool isCommandProhibited(int) = 0;
-    virtual bool createWorkQueue();
-    virtual IO80211WorkQueue *getWorkQueue();
-    virtual void requestPacketTx(void*, UInt);
-    virtual IOCommandGate *getIO80211CommandGate();
-    virtual IOReturn getHardwareAddressForInterface(IOEthernetAddress *);
-    virtual bool useAppleRSNSupplicant(IO80211VirtualInterface *);
-    virtual IO80211SkywalkInterface* getPrimarySkywalkInterface(void);
-    virtual int bpfOutputPacket(OSObject *,UInt,mbuf_t m);
-    virtual SInt32 monitorModeSetEnabled(bool, UInt);
-    virtual SInt32 apple80211_ioctl(IO80211SkywalkInterface *,unsigned long,void *, bool, bool);
-    virtual SInt32 apple80211VirtualRequest(UInt,int,IO80211VirtualInterface *,void *);
-    virtual SInt32 apple80211SkywalkRequest(UInt,int,IO80211SkywalkInterface *,void *);
-    virtual SInt32 apple80211SkywalkRequest(UInt,int,IO80211SkywalkInterface *,void *,void *);
-    
-    virtual SInt32 handleCardSpecific(IO80211SkywalkInterface *,unsigned long,void *,bool) = 0;
-    
-    virtual UInt32 hardwareOutputQueueDepth();
-    virtual SInt32 performCountryCodeOperation(IO80211CountryCodeOp);
-    
-    virtual void dataLinkLayerAttachComplete();
-    virtual SInt32 enableFeature(IO80211FeatureCode, void*) = 0;
-    
-    virtual IOReturn getDRIVER_VERSION(IO80211SkywalkInterface *,apple80211_version_data *) = 0;
-    virtual IOReturn getHARDWARE_VERSION(IO80211SkywalkInterface *,apple80211_version_data *) = 0;
-    virtual IOReturn getCARD_CAPABILITIES(IO80211SkywalkInterface *,apple80211_capability_data *) = 0;
-    virtual IOReturn getPOWER(IO80211SkywalkInterface *,apple80211_power_data *) = 0;
-    virtual IOReturn setPOWER(IO80211SkywalkInterface *,apple80211_power_data *) = 0;
-    virtual IOReturn getCOUNTRY_CODE(IO80211SkywalkInterface *,apple80211_country_code_data *) = 0;
-    virtual IOReturn setCOUNTRY_CODE(IO80211SkywalkInterface *,apple80211_country_code_data *) = 0;
-    virtual IOReturn setGET_DEBUG_INFO(IO80211SkywalkInterface *,apple80211_debug_command *) = 0;
-    
-    virtual SInt32 setVirtualHardwareAddress(IO80211VirtualInterface *,ether_addr *);
-    virtual SInt32 enableVirtualInterface(IO80211VirtualInterface *);
-    virtual SInt32 disableVirtualInterface(IO80211VirtualInterface *);
-    virtual bool requiresExplicitMBufRelease();
-    virtual bool flowIdSupported() {
-        return false;
-    }
-    virtual IO80211FlowQueueLegacy* requestFlowQueue(FlowIdMetadata const*);
-    virtual void releaseFlowQueue(IO80211FlowQueue *);
-    virtual bool getLogPipes(CCPipe**, CCPipe**, CCPipe**);
-    virtual void enableFeatureForLoggingFlags(unsigned long long) {};
-    virtual IOReturn requestQueueSizeAndTimeout(unsigned short *, unsigned short *) { return kIOReturnIOError; };
-    virtual IOReturn enablePacketTimestamping(void) {
-        return kIOReturnUnsupported;
-    }
-    virtual IOReturn disablePacketTimestamping(void) {
-        return kIOReturnUnsupported;
-    }
-    
-    virtual UInt getPacketTSCounter();
-    virtual void *getDriverTextLog();
-    
-    virtual UInt32 selfDiagnosticsReport(int,char const*,UInt);
-    
-    virtual void *getFaultReporterFromDriver();
-    
-    virtual UInt32 getDataQueueDepth(OSObject *);
-    virtual bool isAssociatedToMovingNetwork(void) { return false; }
-    virtual bool wasDynSARInFailSafeMode(void) { return false; }
-    virtual void updateAdvisoryScoresIfNeed(void);
-    virtual UInt64 getAVCAdvisoryInfo(IO80211InterfaceAVCAdvisory *);
-    virtual SInt32 apple80211_ioctl_get(IO80211SkywalkInterface *,void *, bool, bool);
-    virtual SInt32 apple80211_ioctl_set(IO80211SkywalkInterface *,void *, bool, bool);
-    virtual bool attachInterface(OSObject *,IOService *);
-    virtual SInt32 apple80211_ioctl_get(IO80211VirtualInterface *,void *,bool,bool);
-    virtual SInt32 apple80211_ioctl_set(IO80211VirtualInterface *,void *,bool,bool);
-    virtual void detachInterface(OSObject *,bool);
-    virtual IO80211VirtualInterface* createVirtualInterface(ether_addr *,UInt);
-    virtual bool attachVirtualInterface(IO80211VirtualInterface **,ether_addr *,UInt,bool);
-    virtual bool detachVirtualInterface(IO80211VirtualInterface *,bool);
-    virtual IOReturn enable(IO80211SkywalkInterface *);
-    virtual IOReturn disable(IO80211SkywalkInterface *);
-    
-    OSMetaClassDeclareReservedUnused( IO80211Controller,  0);
-    OSMetaClassDeclareReservedUnused( IO80211Controller,  1);
-    OSMetaClassDeclareReservedUnused( IO80211Controller,  2);
-    OSMetaClassDeclareReservedUnused( IO80211Controller,  3);
-    OSMetaClassDeclareReservedUnused( IO80211Controller,  4);
-    OSMetaClassDeclareReservedUnused( IO80211Controller,  5);
-    OSMetaClassDeclareReservedUnused( IO80211Controller,  6);
-    OSMetaClassDeclareReservedUnused( IO80211Controller,  7);
-    OSMetaClassDeclareReservedUnused( IO80211Controller,  8);
-    OSMetaClassDeclareReservedUnused( IO80211Controller,  9);
-    OSMetaClassDeclareReservedUnused( IO80211Controller, 10);
-    OSMetaClassDeclareReservedUnused( IO80211Controller, 11);
-    OSMetaClassDeclareReservedUnused( IO80211Controller, 12);
-    OSMetaClassDeclareReservedUnused( IO80211Controller, 13);
-    OSMetaClassDeclareReservedUnused( IO80211Controller, 14);
-    OSMetaClassDeclareReservedUnused( IO80211Controller, 15);
-    
-    virtual void postMessage(UInt,void *,unsigned long,UInt,void *);
-    virtual IOReturn setMulticastList(ether_addr const*, UInt);
-
-protected:
-    uint8_t  filler[0x128];
-};
-
-#else
 
 class IO80211Controller : public IOEthernetController {
     OSDeclareAbstractStructors(IO80211Controller)
@@ -491,7 +363,6 @@ public:
 protected:
     uint8_t  filler[0x500];
 };
-#endif
 
 // 0x215: 1 byte, length of channel sequence, should be 16
 // 0x21c: channel sequence, should contain 16 elements of length 12, possibly apple80211_channel (but why 16?)
