+++ include/Airport/IO80211ControllerV2.h
@@ -1,0 +1,261 @@
+//
+//  IOSkywalkInterface.h
+//  itlwm
+//
+//  Created by qcwap on 2023/6/7.
+//  Copyright © 2023 钟先耀. All rights reserved.
+//
+
+#ifndef _IO80211CONTROLLER_H
+#define _IO80211CONTROLLER_H
+
+#if defined(KERNEL) && defined(__cplusplus)
+
+#include <Availability.h>
+#include <libkern/version.h>
+
+// This is necessary, because even the latest Xcode does not support properly targeting 11.0.
+#ifndef __IO80211_TARGET
+#error "Please define __IO80211_TARGET to the requested version"
+#endif
+
+#if VERSION_MAJOR > 8
+#define _MODERN_BPF
+#endif
+
+#include <sys/kpi_mbuf.h>
+
+#include <IOKit/network/IOEthernetController.h>
+
+#include <sys/param.h>
+#include <net/bpf.h>
+
+#include "apple80211_ioctl.h"
+#include "IO80211SkywalkInterface.h"
+#include "IO80211WorkLoop.h"
+#include "IO80211WorkQueue.h"
+#include "CCStream.h"
+#include "CCDataPipe.h"
+#include "CCLogPipe.h"
+#include "CCLogStream.h"
+
+#define AUTH_TIMEOUT            15    // seconds
+
+/*! @enum LinkSpeed.
+ @abstract ???.
+ @discussion ???.
+ @constant LINK_SPEED_80211A 54 Mbps
+ @constant LINK_SPEED_80211B 11 Mbps.
+ @constant LINK_SPEED_80211G 54 Mbps.
+ */
+enum {
+    LINK_SPEED_80211A    = 54000000ul,        // 54 Mbps
+    LINK_SPEED_80211B    = 11000000ul,        // 11 Mbps
+    LINK_SPEED_80211G    = 54000000ul,        // 54 Mbps
+    LINK_SPEED_80211N    = 300000000ul,        // 300 Mbps (MCS index 15, 400ns GI, 40 MHz channel)
+};
+
+enum IO80211CountryCodeOp
+{
+    kIO80211CountryCodeReset,                // Reset country code to world wide default, and start
+    // searching for 802.11d beacon
+};
+typedef enum IO80211CountryCodeOp IO80211CountryCodeOp;
+
+enum IO80211SystemPowerState
+{
+    kIO80211SystemPowerStateUnknown,
+    kIO80211SystemPowerStateAwake,
+    kIO80211SystemPowerStateSleeping,
+};
+typedef enum IO80211SystemPowerState IO80211SystemPowerState;
+
+enum IO80211FeatureCode
+{
+    kIO80211Feature80211n = 1,
+};
+typedef enum IO80211FeatureCode IO80211FeatureCode;
+
+
+class IOSkywalkInterface;
+class IO80211ScanManager;
+
+enum scanSource
+{
+    SOURCE_1,
+};
+
+enum joinStatus
+{
+    STATUS_1,
+};
+
+class IO80211Controller;
+class IO80211Interface;
+class IO82110WorkLoop;
+class IO80211VirtualInterface;
+class IO80211ControllerMonitor;
+class CCLogPipe;
+class CCIOReporterLogStream;
+class CCLogStream;
+class IO80211VirtualInterface;
+class IO80211RangingManager;
+class IO80211FlowQueue;
+class IO80211FlowQueueLegacy;
+class FlowIdMetadata;
+class IOReporter;
+class IO80211InfraInterface;
+extern void IO80211VirtualInterfaceNamerRetain();
+
+
+struct apple80211_hostap_state;
+
+struct apple80211_awdl_sync_channel_sequence;
+struct ieee80211_ht_capability_ie;
+struct apple80211_channel_switch_announcement;
+struct apple80211_beacon_period_data;
+struct apple80211_power_debug_sub_info;
+struct apple80211_stat_report;
+struct apple80211_frame_counters;
+struct apple80211_leaky_ap_event;
+struct apple80211_chip_stats;
+struct apple80211_extended_stats;
+struct apple80211_ampdu_stat_report;
+struct apple80211_btCoex_report;
+struct apple80211_cca_report;
+class CCPipe;
+struct apple80211_lteCoex_report;
+
+//typedef int scanSource;
+//typedef int joinStatus;
+//typedef int CCStreamLogLevel;
+typedef IOReturn (*IOCTL_FUNC)(IO80211Controller*, IO80211Interface*, IO80211VirtualInterface*, apple80211req*, bool);
+extern IOCTL_FUNC gGetHandlerTable[];
+extern IOCTL_FUNC gSetHandlerTable[];
+
+class IO80211InterfaceAVCAdvisory;
+
+class IO80211Controller : public IOEthernetController {
+    OSDeclareAbstractStructors(IO80211Controller)
+
+public:
+
+    virtual void free() APPLE_KEXT_OVERRIDE;
+    virtual bool init(OSDictionary *) APPLE_KEXT_OVERRIDE;
+    virtual IOReturn configureReport(IOReportChannelList *,UInt,void *,void *) APPLE_KEXT_OVERRIDE;
+    virtual IOReturn updateReport(IOReportChannelList *,UInt,void *,void *) APPLE_KEXT_OVERRIDE;
+    virtual bool start(IOService *) APPLE_KEXT_OVERRIDE;
+    virtual void stop(IOService *) APPLE_KEXT_OVERRIDE;
+    virtual IOWorkLoop* getWorkLoop(void) const APPLE_KEXT_OVERRIDE;
+    virtual const char* stringFromReturn(int) APPLE_KEXT_OVERRIDE;
+    virtual int errnoFromReturn(int) APPLE_KEXT_OVERRIDE;
+    virtual UInt32 getFeatures() const APPLE_KEXT_OVERRIDE;
+    virtual const OSString * newVendorString() const APPLE_KEXT_OVERRIDE;
+    virtual const OSString * newModelString() const APPLE_KEXT_OVERRIDE;
+    virtual bool createWorkLoop() APPLE_KEXT_OVERRIDE;
+    virtual IOReturn getHardwareAddress(IOEthernetAddress *) APPLE_KEXT_OVERRIDE;
+    virtual IOReturn setHardwareAddress(const IOEthernetAddress * addrP) APPLE_KEXT_OVERRIDE;
+    virtual IOReturn setMulticastMode(bool active) APPLE_KEXT_OVERRIDE;
+    virtual IOReturn setPromiscuousMode(bool active) APPLE_KEXT_OVERRIDE;
+    virtual bool isCommandProhibited(int) = 0;
+    virtual bool createWorkQueue();
+    virtual IO80211WorkQueue *getWorkQueue();
+    virtual void requestPacketTx(void*, UInt);
+    virtual IOCommandGate *getIO80211CommandGate();
+    virtual IOReturn getHardwareAddressForInterface(IOEthernetAddress *);
+    virtual bool useAppleRSNSupplicant(IO80211VirtualInterface *);
+    virtual IO80211SkywalkInterface* getPrimarySkywalkInterface(void);
+    virtual int bpfOutputPacket(OSObject *,UInt,mbuf_t m);
+    virtual SInt32 monitorModeSetEnabled(bool, UInt);
+    virtual SInt32 apple80211_ioctl(IO80211SkywalkInterface *,unsigned long,void *, bool, bool);
+    virtual SInt32 apple80211VirtualRequest(UInt,int,IO80211VirtualInterface *,void *);
+    virtual SInt32 apple80211SkywalkRequest(UInt,int,IO80211SkywalkInterface *,void *);
+    virtual SInt32 apple80211SkywalkRequest(UInt,int,IO80211SkywalkInterface *,void *,void *);
+    
+    virtual SInt32 handleCardSpecific(IO80211SkywalkInterface *,unsigned long,void *,bool) = 0;
+    
+    virtual UInt32 hardwareOutputQueueDepth();
+    virtual SInt32 performCountryCodeOperation(IO80211CountryCodeOp);
+    
+    virtual void dataLinkLayerAttachComplete();
+    virtual SInt32 enableFeature(IO80211FeatureCode, void*) = 0;
+    
+    virtual IOReturn getDRIVER_VERSION(IO80211SkywalkInterface *,apple80211_version_data *) = 0;
+    virtual IOReturn getHARDWARE_VERSION(IO80211SkywalkInterface *,apple80211_version_data *) = 0;
+    virtual IOReturn getCARD_CAPABILITIES(IO80211SkywalkInterface *,apple80211_capability_data *) = 0;
+    virtual IOReturn getPOWER(IO80211SkywalkInterface *,apple80211_power_data *) = 0;
+    virtual IOReturn setPOWER(IO80211SkywalkInterface *,apple80211_power_data *) = 0;
+    virtual IOReturn getCOUNTRY_CODE(IO80211SkywalkInterface *,apple80211_country_code_data *) = 0;
+    virtual IOReturn setCOUNTRY_CODE(IO80211SkywalkInterface *,apple80211_country_code_data *) = 0;
+    virtual IOReturn setGET_DEBUG_INFO(IO80211SkywalkInterface *,apple80211_debug_command *) = 0;
+    
+    virtual SInt32 setVirtualHardwareAddress(IO80211VirtualInterface *,ether_addr *);
+    virtual SInt32 enableVirtualInterface(IO80211VirtualInterface *);
+    virtual SInt32 disableVirtualInterface(IO80211VirtualInterface *);
+    virtual bool requiresExplicitMBufRelease();
+    virtual bool flowIdSupported() {
+        return false;
+    }
+    virtual IO80211FlowQueueLegacy* requestFlowQueue(FlowIdMetadata const*);
+    virtual void releaseFlowQueue(IO80211FlowQueue *);
+    virtual bool getLogPipes(CCPipe**, CCPipe**, CCPipe**);
+    virtual void enableFeatureForLoggingFlags(unsigned long long) {};
+    virtual IOReturn requestQueueSizeAndTimeout(unsigned short *, unsigned short *) { return kIOReturnIOError; };
+    virtual IOReturn enablePacketTimestamping(void) {
+        return kIOReturnUnsupported;
+    }
+    virtual IOReturn disablePacketTimestamping(void) {
+        return kIOReturnUnsupported;
+    }
+    
+    virtual UInt getPacketTSCounter();
+    virtual void *getDriverTextLog();
+    
+    virtual UInt32 selfDiagnosticsReport(int,char const*,UInt);
+    
+    virtual void *getFaultReporterFromDriver();
+    
+    virtual UInt32 getDataQueueDepth(OSObject *);
+    virtual bool isAssociatedToMovingNetwork(void) { return false; }
+    virtual bool wasDynSARInFailSafeMode(void) { return false; }
+    virtual void updateAdvisoryScoresIfNeed(void);
+    virtual UInt64 getAVCAdvisoryInfo(IO80211InterfaceAVCAdvisory *);
+    virtual SInt32 apple80211_ioctl_get(IO80211SkywalkInterface *,void *, bool, bool);
+    virtual SInt32 apple80211_ioctl_set(IO80211SkywalkInterface *,void *, bool, bool);
+    virtual bool attachInterface(OSObject *,IOService *);
+    virtual SInt32 apple80211_ioctl_get(IO80211VirtualInterface *,void *,bool,bool);
+    virtual SInt32 apple80211_ioctl_set(IO80211VirtualInterface *,void *,bool,bool);
+    virtual void detachInterface(OSObject *,bool);
+    virtual IO80211VirtualInterface* createVirtualInterface(ether_addr *,UInt);
+    virtual bool attachVirtualInterface(IO80211VirtualInterface **,ether_addr *,UInt,bool);
+    virtual bool detachVirtualInterface(IO80211VirtualInterface *,bool);
+    virtual IOReturn enable(IO80211SkywalkInterface *);
+    virtual IOReturn disable(IO80211SkywalkInterface *);
+    
+    OSMetaClassDeclareReservedUnused( IO80211Controller,  0);
+    OSMetaClassDeclareReservedUnused( IO80211Controller,  1);
+    OSMetaClassDeclareReservedUnused( IO80211Controller,  2);
+    OSMetaClassDeclareReservedUnused( IO80211Controller,  3);
+    OSMetaClassDeclareReservedUnused( IO80211Controller,  4);
+    OSMetaClassDeclareReservedUnused( IO80211Controller,  5);
+    OSMetaClassDeclareReservedUnused( IO80211Controller,  6);
+    OSMetaClassDeclareReservedUnused( IO80211Controller,  7);
+    OSMetaClassDeclareReservedUnused( IO80211Controller,  8);
+    OSMetaClassDeclareReservedUnused( IO80211Controller,  9);
+    OSMetaClassDeclareReservedUnused( IO80211Controller, 10);
+    OSMetaClassDeclareReservedUnused( IO80211Controller, 11);
+    OSMetaClassDeclareReservedUnused( IO80211Controller, 12);
+    OSMetaClassDeclareReservedUnused( IO80211Controller, 13);
+    OSMetaClassDeclareReservedUnused( IO80211Controller, 14);
+    OSMetaClassDeclareReservedUnused( IO80211Controller, 15);
+    
+    virtual void postMessage(UInt,void *,unsigned long,UInt,void *);
+    virtual IOReturn setMulticastList(ether_addr const*, UInt);
+
+protected:
+    uint8_t  filler[0x128];
+};
+
+#endif /* defined(KERNEL) && defined(__cplusplus) */
+
+#endif /* !_IO80211CONTROLLER_H */
